version 1.0 cleaned some bug fixed

+++ src/beans/validation/CommonUtilisateurBean.java
@@ -1,10 +1,12 @@
 package beans.validation;
 
 
+import java.io.File;
 import java.io.IOException;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 import javax.annotation.PostConstruct;
 import javax.ejb.EJB;
@@ -14,7 +16,12 @@ import javax.faces.bean.ViewScoped;
 import javax.faces.context.ExternalContext;
 import javax.faces.context.FacesContext;
 import javax.servlet.http.HttpSession;
+
+import beans.persistent.Album;
+import beans.persistent.Photo;
 import beans.persistent.Utilisateur;
+import dao.AlbumDao;
+import dao.PhotoDao;
 import dao.UtilisateurDao;
 
 
@@ -25,6 +32,12 @@ public class CommonUtilisateurBean implements Serializable {
 
     @EJB
     UtilisateurDao utilisateurDao;
+    
+    @EJB
+    AlbumDao albumDao;
+    
+    @EJB
+    PhotoDao photoDao;
     
     private List<Utilisateur> utilisateurs;
     private Utilisateur utilisateurProfilDetails = null;
@@ -67,27 +80,58 @@ public class CommonUtilisateurBean implements Serializable {
     	}
     	try {
     		if(isAdminOrSelProfilOrRedirect(username)) {
-        		utilisateurDao.delete(username);
-        		
-            	Utilisateur utilisateurCourant = (Utilisateur)session.getAttribute("utilisateur");
-            	if(utilisateurCourant.getEstAdmin()) {
-            		FacesContext fContext = FacesContext.getCurrentInstance();
-                	ExternalContext extContext = fContext.getExternalContext();
-                	try {
-            			extContext.redirect("list_utilisateur.xhtml");
-            		} catch (IOException e) {
-            			e.printStackTrace();
-            		}
-            	}else {
-            		FacesContext fContext = FacesContext.getCurrentInstance();
-                	ExternalContext extContext = fContext.getExternalContext();
-                	try {
-                		session.removeAttribute("utilisateur");
-            			extContext.redirect("index.xhtml");
-            		} catch (IOException e) {
-            			e.printStackTrace();
-            		}
+    			
+    			Utilisateur user = utilisateurDao.findByUsername(username);
+    			if(user != null) {
+    				List<Album> albumsUtilisateur = albumDao.findByProprietaireUsername(username);
+    				
+    				for (Album a : albumsUtilisateur) {
+    					
+    					List<Photo> photosForAlbum = photoDao.findByAlbumId(a.getId());
+
+    					//Pour chaque album je supprime l'ensemble de ces images
+			    		String photoLocation = FacesContext.getCurrentInstance().getExternalContext().getRealPath("\\resources\\uploaded")+"\\photo_images";
+			            for (Photo p : photosForAlbum) {
+			            	String oldFileName = p.getUri();
+				            File oldFile = new File(photoLocation+"\\"+oldFileName);
+				            photoDao.remove(p.getId());
+				            oldFile.delete();
+						}
+    			            
+    		        	//puis je supprime l'album
+			            String albumLocation = FacesContext.getCurrentInstance().getExternalContext().getRealPath("\\resources\\uploaded")+"\\album_images";
+			    		String oldFileName = a.getUri();
+			            File oldFile = new File(albumLocation+"\\"+oldFileName);
+			            albumDao.remove(a.getId());
+			            oldFile.delete();
+    		    		
+					}
+    				
+    				//Une fois l'ensemble des albums et photo de l'utilisateur supprimé je supprime l'utilisateur
+	        		utilisateurDao.delete(username);
+	        		
+	            	Utilisateur utilisateurCourant = (Utilisateur)session.getAttribute("utilisateur");
+	            	if(utilisateurCourant.getEstAdmin()) {
+	            		FacesContext fContext = FacesContext.getCurrentInstance();
+	                	ExternalContext extContext = fContext.getExternalContext();
+	                	try {
+	            			extContext.redirect("list_utilisateur.xhtml");
+	            		} catch (IOException e) {
+	            			e.printStackTrace();
+	            		}
+	            	}else {
+	            		FacesContext fContext = FacesContext.getCurrentInstance();
+	                	ExternalContext extContext = fContext.getExternalContext();
+	                	try {
+	                		session.removeAttribute("utilisateur");
+	            			extContext.redirect("index.xhtml");
+	            		} catch (IOException e) {
+	            			e.printStackTrace();
+	            		}
+	            	}
+            	
             	}
+            	
         	}
 		} catch (Exception e) {
 			FacesContext.getCurrentInstance().addMessage(null, new FacesMessage(
